Composition

+++ OOPS-ShapesProject/Quad.cs
@@ -6,12 +6,12 @@ namespace OOPS_ShapesProject {
 
     public class Quad {
 
-        public int side1 { get; set; }
-        public int side2 { get; set; }
-        public int side3 { get; set; }
-        public int side4 { get; set; }
+        public double side1 { get; set; }
+        public double side2 { get; set; }
+        public double side3 { get; set; }
+        public double side4 { get; set; }
 
-        public int GetPerimeter() {
+        public double GetPerimeter() {
 
           return side1 + side2 + side3 + side4;
